app now has default sound, no need to go to settings once before sound starts working
