+++ app/build.gradle
@@ -40,6 +40,9 @@ android {
         targetCompatibility 1.8
         sourceCompatibility 1.8
     }
+    buildFeatures {
+        viewBinding true
+    }
 }
 
 dependencies {
@@ -49,16 +52,20 @@ dependencies {
     })
     implementation 'com.journeyapps:zxing-android-embedded:3.6.0'
     implementation 'androidx.appcompat:appcompat:1.6.1'
-    implementation 'com.google.android.material:material:1.10.0'
+    implementation 'com.google.android.material:material:1.11.0'
     implementation 'androidx.constraintlayout:constraintlayout:2.1.4'
     implementation 'androidx.legacy:legacy-support-v4:1.0.0'
     implementation 'androidx.vectordrawable:vectordrawable:1.1.0'
     implementation 'com.github.apl-devs:appintro:v4.2.0'
+    implementation "androidx.navigation:navigation-fragment-ktx:2.7.7"
+    implementation 'androidx.preference:preference-ktx:1.1.1'
+    implementation "androidx.navigation:navigation-ui-ktx:2.7.7"
     testImplementation 'junit:junit:4.13.2'
     implementation files('libs/poi-3.12-android-a.jar')
+    implementation("androidx.navigation:navigation-fragment-ktx:2.7.7")
     implementation files('libs/poi-ooxml-schemas-3.12-20150511-a.jar')
     implementation "org.jetbrains.kotlin:kotlin-stdlib-jdk7:$kotlin_version"
-    implementation 'com.google.android.material:material:1.9.0'
+    implementation 'com.google.android.material:material:1.11.0'
     implementation 'com.github.daniel-stoneuk:material-about-library:3.1.2'
 }
 repositories {
