+++ app/src/main/java/org/phenoapps/verify/LoaderDBActivity.java
@@ -6,13 +6,9 @@ import android.database.sqlite.SQLiteDatabase;
 import android.database.sqlite.SQLiteException;
 import android.net.Uri;
 import android.os.Bundle;
-
-import androidx.appcompat.widget.ActionMenuView;
 import androidx.core.app.ActivityCompat;
 import androidx.appcompat.app.AppCompatActivity;
 import android.util.Log;
-import android.view.Menu;
-import android.view.MenuInflater;
 import android.view.MenuItem;
 import android.view.View;
 import android.widget.AbsListView;
@@ -75,7 +71,7 @@ public class LoaderDBActivity extends AppCompatActivity {
         setContentView(R.layout.activity_load_file);
 
         if(getSupportActionBar() != null){
-            getSupportActionBar().setTitle(null);
+            getSupportActionBar().setTitle("Import Data");
             getSupportActionBar().getThemedContext();
             getSupportActionBar().setDisplayHomeAsUpEnabled(true);
             getSupportActionBar().setHomeButtonEnabled(true);
@@ -99,8 +95,8 @@ public class LoaderDBActivity extends AppCompatActivity {
 
         mFileUri = getIntent().getData();
 
-        if (mFileUri == null ) {
-            Toast.makeText(this, getString(R.string.problem_reading_file), Toast.LENGTH_LONG).show();
+        if (mFileUri == null ){
+            Toast.makeText(this, "There was a problem reading this file", Toast.LENGTH_LONG).show();
             return;
         }
 
@@ -113,7 +109,7 @@ public class LoaderDBActivity extends AppCompatActivity {
 
         //if unsupported file type, start delimiter tutorial
         if (mDelimiter == null || mHeader == null) {
-            Toast.makeText(this, getString(R.string.problem_reading_file_general), Toast.LENGTH_LONG).show();
+            Toast.makeText(this, "There was a problem reading this file.", Toast.LENGTH_LONG).show();
             finish();
            /* if (mHeader == null) {
                 tutorialText.setText("Error reading file.");
@@ -139,7 +135,7 @@ public class LoaderDBActivity extends AppCompatActivity {
             int lastDot = mFilePath.lastIndexOf("."); // changed from mFileUri to mFilePath due to the files in download folder have URI without extension
 
             if (lastDot == -1) {
-                Toast.makeText(this, getString(R.string.must_have_extension), Toast.LENGTH_LONG).show();
+                Toast.makeText(this, "Imported file must have an extension. (e.g: .csv, .tsv)", Toast.LENGTH_LONG).show();
                 finish();
             }
 
@@ -493,7 +489,7 @@ public class LoaderDBActivity extends AppCompatActivity {
 
         if (mHeader == null) {
             headerList.setAdapter(new ArrayAdapter<String>(this, org.phenoapps.verify.R.layout.row));
-            tutorialText.setText(getString(R.string.error_reading_file));
+            tutorialText.setText("Error reading file.");
             return;
         }
 
@@ -507,7 +503,7 @@ public class LoaderDBActivity extends AppCompatActivity {
             headerList.setAdapter(idAdapter);
         } else {
             headerList.setAdapter(new ArrayAdapter<String>(this, org.phenoapps.verify.R.layout.row));
-            tutorialText.setText(getString(R.string.error_reading_file));
+            tutorialText.setText("Error reading file.");
         }
     }
 
@@ -517,26 +513,6 @@ public class LoaderDBActivity extends AppCompatActivity {
 
         if (mFileUri != null)
             outState.putString(VerifyConstants.CSV_URI, mFileUri.toString());
-    }
-    @Override
-    final public boolean onCreateOptionsMenu(Menu m) {
-
-        final MenuInflater inflater = getMenuInflater();
-        inflater.inflate(org.phenoapps.verify.R.menu.activity_main_toolbar, m);
-
-        ActionMenuView bottomToolBar = (ActionMenuView) findViewById(R.id.bottom_toolbar);
-        Menu bottomMenu = bottomToolBar.getMenu();
-        inflater.inflate(R.menu.activity_main_bottom_toolbar, bottomMenu);
-
-        for (int i = 0; i < bottomMenu.size(); i++) {
-            bottomMenu.getItem(i).setOnMenuItemClickListener(new MenuItem.OnMenuItemClickListener() {
-                @Override
-                public boolean onMenuItemClick(MenuItem item) {
-                    return onOptionsItemSelected(item);
-                }
-            });
-        }
-        return true;
     }
 
     @Override
