+++ app/src/main/java/org/phenoapps/verify/SettingsFragment.java
@@ -5,14 +5,28 @@ import android.content.Intent;
 import android.content.SharedPreferences;
 import android.os.Build;
 import android.os.Bundle;
-import android.preference.ListPreference;
-import android.preference.Preference;
-import android.preference.PreferenceFragment;
 import android.widget.Toast;
+
+import androidx.preference.Preference;
+import androidx.preference.ListPreference;
+import androidx.preference.PreferenceFragmentCompat;
 
 import java.util.prefs.Preferences;
 
-public class SettingsFragment extends PreferenceFragment {
+public class SettingsFragment extends PreferenceFragmentCompat {
+
+
+    public static final CharSequence INTRO_BUTTON = "org.phenoapps.verify.INTRO";
+    public static final CharSequence ABOUT_BUTTON = "org.phenoapps.verify.ABOUT";
+    public static String FILE_NAME = "org.phenoapps.verify.FILE_NAME";
+    public static String SCAN_MODE_LIST = "org.phenoapps.verify.SCAN_MODE";
+    public static String AUDIO_ENABLED = "org.phenoapps.verify.AUDIO_ENABLED";
+    public static String TUTORIAL_MODE = "org.phenoapps.verify.TUTORIAL_MODE";
+    public static String NAME = "org.phenoapps.verify.NAME";
+    public static String LIST_KEY_NAME = "org.phenoapps.verify.LIST_KEY_NAME";
+    public static String PAIR_NAME = "org.phenoapps.verify.PAIR_NAME";
+    public static String DISABLE_PAIR = "org.phenoapps.verify.DISABLE_PAIR";
+    public static String AUX_INFO = "org.phenoapps.verify.AUX_INFO";
 
 
     private void showChangeLog() {
@@ -30,12 +44,18 @@ public class SettingsFragment extends PreferenceFragment {
     public void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
 
-        addPreferencesFromResource(org.phenoapps.verify.R.xml.preferences);
+    }
+
+    @Override
+    public void onCreatePreferences(Bundle savedInstanceState, String rootKey) {
+
+        setPreferencesFromResource(R.xml.preferences, rootKey);
+
 
         final SharedPreferences sharedPrefs = super.getPreferenceManager().getSharedPreferences();
-        ListPreference mode = (ListPreference) findPreference(SettingsActivity.SCAN_MODE_LIST);
-        Preference introButton = (Preference) findPreference(SettingsActivity.INTRO_BUTTON);
-        Preference aboutButton = (Preference) findPreference(SettingsActivity.ABOUT_BUTTON);
+        ListPreference mode = findPreference(SCAN_MODE_LIST);
+        Preference introButton = findPreference(INTRO_BUTTON);
+        Preference aboutButton = findPreference(ABOUT_BUTTON);
 
         aboutButton.setOnPreferenceClickListener(new Preference.OnPreferenceClickListener() {
             @Override
@@ -68,7 +88,7 @@ public class SettingsFragment extends PreferenceFragment {
             @Override
             public boolean onPreferenceChange(Preference preference, Object o) {
                 if (o.equals("4") &&
-                        sharedPrefs.getBoolean(SettingsActivity.DISABLE_PAIR, false)) {
+                        sharedPrefs.getBoolean(DISABLE_PAIR, false)) {
                     ((ListPreference) preference).setValue("0");
                     Toast.makeText(getActivity(),
                             "Pair mode cannot be used without setting a pair ID.",
