+++ app/src/main/res/layout/activity_main.xml
@@ -1,133 +1,27 @@
-<?xml version="1.0" encoding="utf-8"?>
-<androidx.drawerlayout.widget.DrawerLayout xmlns:android="http://schemas.android.com/apk/res/android"
+<androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
-    android:id="@+id/drawer_layout"
     android:layout_width="match_parent"
     android:layout_height="match_parent">
 
-    <androidx.constraintlayout.widget.ConstraintLayout
-        android:id="@+id/constraint_layout_parent"
+    <fragment
+        android:id="@+id/nav_host_fragment"
+        android:background="?android:windowBackground"
+        android:clickable = "true"
+        android:name="androidx.navigation.fragment.NavHostFragment"
+        app:navGraph="@navigation/mobile_navigation"
         android:layout_width="match_parent"
-        android:layout_height="match_parent">
+        android:layout_height="match_parent"
+        android:focusable="true" />
 
-        <ListView
-            android:id="@+id/idTable"
-            android:layout_width="0dp"
-            android:layout_height="200dp"
-                        android:visibility="visible"
-            app:layout_constraintLeft_toLeftOf="parent"
-            app:layout_constraintRight_toRightOf="parent"
-            app:layout_constraintTop_toTopOf="parent" />
-
-        <EditText
-            android:id="@+id/scannerTextView"
-            style="@style/Base.TextAppearance.AppCompat.Headline"
-            android:layout_width="0dp"
-            android:layout_height="60dp"
-            android:layout_margin="8dp"
-            android:focusable="true"
-            android:hapticFeedbackEnabled="true"
-            android:inputType="text"
-            android:maxLines="1"
-            android:textAlignment="center"
-            app:layout_constraintLeft_toLeftOf="parent"
-            app:layout_constraintRight_toRightOf="parent"
-            app:layout_constraintTop_toBottomOf="@+id/idTable" />
-
-        <Button
-            android:id="@+id/clearButton"
-            android:layout_width="32dp"
-            android:layout_height="32dp"
-            android:layout_alignParentEnd="true"
-            android:layout_alignParentRight="true"
-            android:layout_centerVertical="true"
-            android:layout_margin="8dp"
-            android:background="@drawable/ic_clear"
-            app:layout_constraintBottom_toBottomOf="@+id/scannerTextView"
-            app:layout_constraintRight_toRightOf="parent"
-            app:layout_constraintTop_toTopOf="@+id/scannerTextView" />
-
-        <LinearLayout
-            android:id="@+id/linearLayout"
-            android:layout_width="0dp"
-            android:layout_height="0dp"
-            android:orientation="vertical"
-            app:layout_constraintBottom_toBottomOf="parent"
-            app:layout_constraintEnd_toEndOf="parent"
-            app:layout_constraintStart_toStartOf="parent"
-            app:layout_constraintTop_toBottomOf="@+id/scannerTextView">
-
-            <ScrollView xmlns:android="http://schemas.android.com/apk/res/android"
-                android:id="@+id/scrollView"
-                android:layout_width="match_parent"
-                android:layout_height="wrap_content"
-                android:layout_weight="0.5"
-                android:fillViewport="true"
-                android:visibility="visible"
-                app:layout_constraintEnd_toEndOf="parent"
-                app:layout_constraintStart_toStartOf="parent"
-                app:layout_constraintTop_toTopOf="parent">
-
-                <TextView
-                    android:id="@+id/valueView"
-                    style="@style/Base.TextAppearance.AppCompat.Title"
-                    android:layout_width="match_parent"
-                    android:layout_height="wrap_content"
-                    android:visibility="visible" />
-
-            </ScrollView>
-
-            <View
-                android:id="@+id/view"
-                android:layout_width="0dp"
-                android:layout_height="0dp"
-                android:background="#66eec657"
-                app:layout_constraintBottom_toBottomOf="parent"
-                app:layout_constraintEnd_toEndOf="parent"
-                app:layout_constraintStart_toStartOf="parent"
-                app:layout_constraintTop_toTopOf="parent" />
-
-           <ScrollView xmlns:android="http://schemas.android.com/apk/res/android"
-               android:id="@+id/auxScrollView"
-               android:layout_width="match_parent"
-               android:layout_height="wrap_content"
-               android:layout_weight="0.5"
-               android:fillViewport="true"
-               android:visibility="visible"
-               app:layout_constraintEnd_toEndOf="parent"
-               app:layout_constraintStart_toStartOf="parent"
-               app:layout_constraintTop_toTopOf="@+id/scrollView">
-
-               <TextView
-               android:id="@+id/auxValueView"
-               style="@style/Base.TextAppearance.AppCompat.Title"
-               android:layout_width="match_parent"
-               android:layout_height="wrap_content"
-               android:visibility="visible" />
-
-           </ScrollView>
-        </LinearLayout>
-
-        <LinearLayout
-            android:layout_width="match_parent"
-            android:layout_height="?attr/actionBarSize"
-            android:background="?attr/colorPrimary"
-            android:gravity="center"
-            android:weightSum="2"
-            app:layout_constraintBottom_toBottomOf="@+id/linearLayout">
-
-            <androidx.appcompat.widget.ActionMenuView
-                android:id="@+id/bottom_toolbar"
-                android:layout_weight="1.5"
-                android:layout_width="0dp"
-                android:layout_height="?attr/actionBarSize"
-                android:gravity="center"
-                android:elevation="4dp"
-                android:foregroundGravity="center"
-                />
-        </LinearLayout>
-
-    </androidx.constraintlayout.widget.ConstraintLayout>
-
-</androidx.drawerlayout.widget.DrawerLayout>
+    <com.google.android.material.bottomnavigation.BottomNavigationView
+        android:id="@+id/bottom_toolbar"
+        android:layout_width="match_parent"
+        android:layout_height="?attr/actionBarSize"
+        app:labelVisibilityMode="labeled"
+        android:background="@color/colorPrimary"
+        app:itemIconTint="@drawable/bottom_nav_color"
+        app:itemTextColor="@drawable/bottom_nav_color"
+        app:menu="@menu/activity_main_bottom_toolbar"
+        app:layout_constraintBottom_toBottomOf="@+id/nav_host_fragment" />
+</androidx.constraintlayout.widget.ConstraintLayout>
