+++ app/src/main/res/navigation/mobile_navigation.xml
@@ -1,0 +1,25 @@
+<?xml version="1.0" encoding="utf-8"?>
+<navigation xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    xmlns:tools="http://schemas.android.com/tools"
+    android:id="@+id/mobile_navigation"
+    app:startDestination="@+id/Home">
+
+    <fragment
+        android:id="@+id/Home"
+        android:name="org.phenoapps.verify.HomeFragment"
+        android:label="Home"
+        tools:layout="@layout/fragment_home" />
+
+    <fragment
+        android:id="@+id/Compare"
+        android:name="org.phenoapps.verify.CompareFragment"
+        android:label="Compare"
+        tools:layout="@layout/activity_compare" />
+
+    <fragment
+        android:id="@+id/Settings"
+        android:name="org.phenoapps.verify.SettingsFragment"
+        android:label="Settings"
+        />
+</navigation>
