+++ app/src/main/res/values-night/themes.xml
@@ -1,0 +1,7 @@
+<resources>
+
+    <style name="ThemeOverlay.B_A_N.FullscreenContainer" parent="">
+        <item name="fullscreenBackgroundColor">@color/light_blue_900</item>
+        <item name="fullscreenTextColor">@color/light_blue_A400</item>
+    </style>
+</resources>
