+++ app/src/main/res/values/colors.xml
@@ -3,4 +3,10 @@
     <color name="colorPrimary">#E4624C</color>
     <color name="colorPrimaryDark">#A84937</color>
     <color name="colorAccent">#EEC657</color>
+    <color name="light_blue_600">#FF039BE5</color>
+    <color name="light_blue_900">#FF01579B</color>
+    <color name="light_blue_A200">#FF40C4FF</color>
+    <color name="light_blue_A400">#FF00B0FF</color>
+    <color name="black_overlay">#66000000</color>
+    <color name="white">#FFFFFF</color>
 </resources>
