+++ app/src/main/res/values/themes.xml
@@ -1,0 +1,7 @@
+<resources>
+
+    <style name="ThemeOverlay.B_A_N.FullscreenContainer" parent="">
+        <item name="fullscreenBackgroundColor">@color/light_blue_600</item>
+        <item name="fullscreenTextColor">@color/light_blue_A200</item>
+    </style>
+</resources>
